+++ crates/module-system/sov-state/src/nomt/zk_storage.rs
@@ -41,6 +41,10 @@ impl<S: MerkleProofSpec> NomtVerifierStorage<S> {
             ordered_writes: state_writes,
         } = state_accesses;
 
+        if state_reads.is_empty() && state_writes.is_empty() {
+            return Ok(prev_root);
+        }
+
         let multi_proof: MultiProof = array_witness.get_hint();
         let verified_multi_proof = nomt_core::proof::verify_multi_proof::<BinaryHasher<S::Hasher>>(
             &multi_proof,
@@ -135,7 +139,7 @@ impl<S: MerkleProofSpec> Storage for NomtVerifierStorage<S> {
     }
 
     fn get_accessory(&self, _key: &SlotKey) -> Option<SlotValue> {
-        unimplemented!("The NomtZkStorage does not have the accessory state yet.")
+        unimplemented!("The NomtVerifierStorage does not have the accessory state yet.")
     }
 
     fn compute_state_update(
@@ -168,8 +172,8 @@ impl<S: MerkleProofSpec> Storage for NomtVerifierStorage<S> {
 }
 
 #[cfg(all(feature = "test-utils", feature = "native"))]
-// `NativeStorage`` is implemented for `ZkStorage` solely for testing purposes.
-// In some tests, we use both `ProverStorage`` and `ZkStorage`.
+// `NativeStorage` is implemented for `NomtVerifierStorage` solely for testing purposes.
+// In some tests, we use both `NomtProverStorage` and `NomtVerifierStorage`.
 // Due to feature unification, we must provide this implementation even though it is not used.
 impl<S: MerkleProofSpec> crate::storage::NativeStorage for NomtVerifierStorage<S> {
     fn latest_version(&self) -> SlotNumber {